Update PHP Return types to allow null returns
Previously, return types had been removed in some areas due to the
fact that adding a return type would not allow null returns. This
has since been added in php 7.1, so the nullable types have been added
in this change.
Formatting and tests were also updated, as well as a couple bugs.