Add Schedule Authorization.
Add admin user list to the schedule object and enforce that only admins
can create events for a schedule.
Introduces a param converter to load schedule objects. this is needed in
order for security annotations to work, and also convienence.
Also adds a voter that currently grants all permissions to any user
in the Schedule's admin list.